refactored ZuulService to encrypte SettingsEntry instead of just string values

+++ zuul-data/src/main/groovy/org/devnull/zuul/data/dao/SettingsGroupDao.groovy
@@ -4,7 +4,7 @@ import org.devnull.zuul.data.model.Environment
 import org.devnull.zuul.data.model.SettingsGroup
 import org.springframework.data.repository.PagingAndSortingRepository
 
-interface SettingsGroupDao extends PagingAndSortingRepository<SettingsGroup, Long> {
+interface SettingsGroupDao extends PagingAndSortingRepository<SettingsGroup, Integer> {
     SettingsGroup findByNameAndEnvironment(String name, Environment environment)
     List<SettingsGroup> findByName(String name)
 }
